Add support for decimal in absolute.py.

+++ absolute.py
@@ -1,9 +1,9 @@
-numberIn=int(raw_input("Type the input number"))
-numberOut=0
+numberIn = float(raw_input("Type the input number"))
+numberOut = 0.0
 
-if numberIn<0 :
-    numberOut=-numberIn
+if numberIn < 0.0:
+    numberOut = -numberIn
 else:
-    numberOut=numberIn
+    numberOut = numberIn
 
 print numberOut
